mod time
- move time conversion function next to each other
- todo about bugs in `system_time_from_time`
