+++ src/ll/reply.rs
@@ -1,10 +1,9 @@
-use std::convert::TryInto;
 use std::io::IoSlice;
 use std::os::unix::prelude::OsStrExt;
 use std::path::Path;
 use std::time::Duration;
+#[cfg(target_os = "macos")]
 use std::time::SystemTime;
-use std::time::UNIX_EPOCH;
 
 use smallvec::SmallVec;
 use smallvec::smallvec;
@@ -21,6 +20,7 @@ use crate::ll::Lock;
 use crate::ll::RequestId;
 use crate::ll::flags::fopen_flags::FopenFlags;
 use crate::ll::fuse_abi as abi;
+use crate::time::time_from_system_time;
 
 const INLINE_DATA_THRESHOLD: usize = size_of::<u64>() * 4;
 pub(crate) type ResponseBuf = SmallVec<[u8; INLINE_DATA_THRESHOLD]>;
@@ -261,30 +261,6 @@ impl<'a> Response<'a> {
     }
 }
 
-pub(crate) fn time_from_system_time(system_time: &SystemTime) -> (i64, u32) {
-    // Convert to signed 64-bit time with epoch at 0
-    match system_time.duration_since(UNIX_EPOCH) {
-        Ok(duration) => match i64::try_from(duration.as_secs()) {
-            Ok(secs) => (secs, duration.subsec_nanos()),
-            Err(_) => (i64::MAX, 999_999_999),
-        },
-        Err(before_epoch_error) => {
-            let d = before_epoch_error.duration();
-            let secs = d.as_secs();
-            let nanos = d.subsec_nanos();
-
-            // Minus min representable value.
-            if (secs, nanos) >= (i64::MAX as u64 + 1, 0) {
-                // Saturate.
-                (i64::MIN, 0)
-            } else if nanos == 0 {
-                (-(secs as i64), 0)
-            } else {
-                (-(secs as i64) - 1, 1_000_000_000 - nanos)
-            }
-        }
-    }
-}
 // Some platforms like Linux x86_64 have mode_t = u32, and lint warns of a trivial_numeric_casts.
 // But others like macOS x86_64 have mode_t = u16, requiring a typecast.  So, just silence lint.
 #[allow(trivial_numeric_casts)]
@@ -521,6 +497,7 @@ impl DirEntPlusList {
 #[cfg(test)]
 mod test {
     use std::num::NonZeroI32;
+    use std::time::UNIX_EPOCH;
 
     use crate::ll::reply::*;
     use crate::ll::test::ioslice_to_vec;
@@ -897,28 +874,5 @@ mod test {
             r.with_iovec(RequestId(0xdeadbeef), ioslice_to_vec),
             expected
         );
-    }
-
-    #[test]
-    fn test_time_from_system_time_negative() {
-        let before_epoch = UNIX_EPOCH - Duration::new(1, 200_000_000);
-        let (secs, nanos) = time_from_system_time(&before_epoch);
-        assert_eq!((-2, 800_000_000), (secs, nanos));
-    }
-
-    #[test]
-    fn test_time_from_system_time_i64_min_boundary() {
-        // timespec { tv_sec: i64::MIN, tv_nsec: 0 }
-        let min_system_time = UNIX_EPOCH - Duration::new(i64::MAX as u64 + 1, 0);
-        let (secs, nanos) = time_from_system_time(&min_system_time);
-        assert_eq!((i64::MIN, 0), (secs, nanos));
-
-        let min_system_time_plus_eps = UNIX_EPOCH - Duration::new(i64::MAX as u64, 800_000_000);
-        let (secs, nanos) = time_from_system_time(&min_system_time_plus_eps);
-        assert_eq!((i64::MIN, 200_000_000), (secs, nanos));
-
-        let min_system_time_plus_one = UNIX_EPOCH - Duration::new(i64::MAX as u64, 0);
-        let (secs, nanos) = time_from_system_time(&min_system_time_plus_one);
-        assert_eq!((i64::MIN + 1, 0), (secs, nanos));
     }
 }
